cast jax array to int

+++ desc/integrals/quad_utils.py
@@ -520,7 +520,7 @@ def _vanilla_params(grid):
     """
     Nt = grid.num_theta
     Nz = grid.num_zeta * grid.NFP
-    q = jnp.sqrt(grid.num_nodes) // 2
+    q = int(jnp.sqrt(grid.num_nodes) // 2)
     s = min(q, Nt, Nz)
     return s, s, q
 
@@ -555,7 +555,7 @@ def _best_params(grid, ratio):
     assert grid.can_fft2
     Nt = grid.num_theta
     Nz = grid.num_zeta * grid.NFP
-    q = jnp.sqrt(grid.num_nodes if (grid.num_zeta > 1) else (Nt * Nz)) // 2
+    q = int(jnp.sqrt(grid.num_nodes if (grid.num_zeta > 1) else (Nt * Nz)) // 2)
     s = min(q, Nt, Nz)
     # Size of singular region in real space = s * h * |e_.|
     # For it to be a circle, choose radius ~ equal
